requirements_extra.txt
plotly>=4.4.1
matplotlib>=3.2.2
pyvista>=0.31
Sphinx>=4.1
numpydoc>=1.1
ipython>=5.5.0
pydata-sphinx-theme>=0.6.3
pytest>=3.6.4
